Extract users to plugin
